requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
boto3==1.29.6
pypdf==3.17.1
numpy<2.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai>=1.10.0
requests==2.31.0
httpx>=0.24.0
chromadb==0.4.18